fix: 시맨틱 태그 변경 및 스크롤 위치 변경, 응답 타입 파일 생성, 훅 파일 이름 변경경
