+++ src/apis/services/user.ts
@@ -3,7 +3,9 @@ import { apiClient } from "../client/APIClient";
 import API_ROUTES from "../constants/routes";
 
 export async function getUserInfo(): Promise<UserType> {
-  return await apiClient.get<UserType>(API_ROUTES.USER);
+  return await apiClient.get<UserType>(API_ROUTES.USER, {
+    skipTokenVerification: true,
+  });
 }
 
 export async function logout() {
