+++ src/app/my/project/all/page.tsx
@@ -4,11 +4,11 @@ import TabWrapper from "@/components/user/project/tab/TabWrapper";
 
 async function ProjectTabPage() {
   return (
-    <main className="w-full px-23 py-28">
+    <section className="w-full px-23 py-28">
       <TabWrapper activeTab="all">
         <TabAllContainer />
       </TabWrapper>
-    </main>
+    </section>
   );
 }
