+++ src/app/my/project/folder/page.tsx
@@ -3,11 +3,11 @@ import TabWrapper from "@/components/user/project/tab/TabWrapper";
 
 async function ProjectFolderTabPage() {
   return (
-    <main className="w-full px-23 py-28">
+    <section className="w-full px-23 py-28">
       <TabWrapper activeTab="folder">
         <TabFolderContainer />
       </TabWrapper>
-    </main>
+    </section>
   );
 }
