+++ src/hooks/home/useGetImageList.ts
@@ -1,18 +1,12 @@
-import { ImageType, PageNation } from "@/@types/images";
 import { getSharedImages } from "@/apis/services/images";
 import { QUERY_KEY } from "@/constants/query-key";
 import { useInfiniteQuery } from "@tanstack/react-query";
 
 export default function useGetImageList() {
-  return useInfiniteQuery<
-    PageNation<ImageType>,
-    Error,
-    PageNation<ImageType>,
-    typeof QUERY_KEY.IMAGE.LIST,
-    string | null
-  >({
+  return useInfiniteQuery({
     queryKey: QUERY_KEY.IMAGE.LIST,
-    queryFn: ({ pageParam = null }) => getSharedImages({ cursor: pageParam }),
+    queryFn: ({ pageParam = null }: { pageParam?: string | null }) =>
+      getSharedImages({ cursor: pageParam }),
     initialPageParam: null,
     getNextPageParam: (lastPage) => lastPage.nextPageCursor ?? undefined,
     getPreviousPageParam: (firstPage) =>
